+++ fragments/default_theme/mform_select.php
@@ -4,6 +4,13 @@ switch ($this->type) {
     case 'multiselect';
         $this->attributes = $this->attributes . ' multiple';
     case 'select';
+        foreach (rex_view::getJsFiles() as $file) {
+            if (str_contains($file, 'bootstrap-select')
+                && !str_contains($this->class, 'selectpicker')
+                && !str_contains($this->class, 'none-selectpicker')) {
+                $this->class = $this->class . ' selectpicker';
+            }
+        }
         echo $this->javascript . '<select id="' . $this->id . '" name="REX_INPUT_VALUE' . $this->varId . '" class="' . $this->class . '" ' . $this->attributes . ' data-selected="' . $this->value . '">' . $this->options . '</select>' . $this->hidden;
         break;
     case 'option':